requirements.txt
opencv-python
python-decouple
imutils
pyautogui